add security 7

+++ src/main/java/com/maxim/taskmanager/controller/UserController.java
@@ -1,14 +1,19 @@
 package com.maxim.taskmanager.controller;
 
+import com.maxim.taskmanager.exception.UserNotFoundException;
 import com.maxim.taskmanager.model.dto.UserDto.UserCreateDto;
 import com.maxim.taskmanager.model.dto.UserDto.UserResponseDto;
 import com.maxim.taskmanager.model.dto.UserDto.UserUpdateDto;
+import com.maxim.taskmanager.model.entity.User;
+import com.maxim.taskmanager.repository.UserRepository;
 import com.maxim.taskmanager.service.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -27,6 +32,7 @@ import java.util.List;
 public class UserController {
 
     private final UserService userService;
+    private final UserRepository userRepository;
 
     @GetMapping("/{id}")
     public ResponseEntity<UserResponseDto> getUserById(@PathVariable Integer id) {
@@ -63,9 +69,21 @@ public class UserController {
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteUser(@PathVariable Integer id) {
         log.info("DELETE /api/users/{} - удаление пользователя", id);
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        String currentUserEmail = auth.getName();
+        User userToDelete = userRepository.findById(id)
+                .orElseThrow(() -> new UserNotFoundException("Пользователь с id " + id + " не найден"));
+        boolean isAdmin = auth.getAuthorities().stream()
+                .anyMatch(granted -> granted.getAuthority().equals("ROLE_ADMIN"));
+        boolean isSelf = userToDelete.getEmail().equals(currentUserEmail);
+
+        if (!isAdmin && !isSelf) {
+            log.warn("Пользователь {} пытается удалить пользователя {}", currentUserEmail, userToDelete.getEmail());
+            throw new RuntimeException("У вас нет прав на удаление этого пользователя");
+        }
         userService.deleteUser(id);
         log.info("DELETE /api/users/{} - пользователь удалён", id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        return ResponseEntity.noContent().build();
     }
 
     @PutMapping("/{id}")
